Make date, size, md5sum and regexp diffable.
This is sane, I think.
Also, the behavior of regexp changed:
 - before commit: matching regexp results in prompted uri
 -  after commit: change of match status results in prompted uri
This is more intuitive and the `diff' command shows whether it matched or not.

A note about the code: my Ruby is a little rusty and because I was
a little unsure about the whole "proc" stuff, I that to "yield" stuff ;-)
I did not want to do any experiments, although the usage of proc
may be more elegant in some cases.