couple of changes to make it run on my system
First the name of the Config class is not available resulting in the following
error: "Config is not a class (TypeError)" (My ruby version is 1.8.6)

The second is a problem mentioned in the linux.com article where the regexp is
not always saved in the config file. This is because in some cases the "test"
method is nil (when the user hasn't specifically made a choice).

I've attached a patch with both changes.

Regards

Konstantinos